move getMultipleHeaderData to avail funnel utils
